Allow storing resume state in DB
This commit introduces the ability to store the state to resume from in
a dedicated database, accessed via the "target DB" handle. This is
useful when the client should not store any state, or when we need to
guarantee that a write to the target DB also updates the state in the
same transaction (avoiding double-applying copy/binlog operations).

Change-Id: Id88bc7dad322676e930306a0153d7339ae5aaf12