Support copying tables without "paging" keys #162 (#6)
This commit adds the ability to mark tables as "full-table copy" tables,
for which we copy all table rows using a simple SQL table "offset",
allowing to copy (small) tables that do not have indices we can use to
iterate on.

To guarantee correct copy behavior, the source table is locked for the
entire copy procedure, making this viable only for small tables.

Change-Id: Ie009175d94fe9d2c4fd75b63e91243e2260b909d