CINF-324 Fix config validation #17
Fix the check if the configuration is compatible with replication of
schema changes if no verifier is specified.

NOTE: After several failures on travis-ci, this conmit also disabled one
of the racy integration tests. They work very reliably on a fast
machine, but on an underpowered VM, they randomly fail.
The test implementation is not great, but it tests a race-conditoin so
it's hard to write it much  better.

Change-Id: I6b5cc94ed82b14ed33d6416cc6b4c413c34aaaee