Allow operating on read-only source DBs
Allow disabling locking for copies: We cannot lock tables on a
read-only DB (unless we have the SUPER privilege).
This is only safe depending on the application and thus off by default.

Move AllowReplicationFromReplica into config: This commit moves the
flag for allowing repliction from a slave DB from the Ferry object
into the Config object to allow enabling it for copydb via the config.

This commit fixes #21 and fixes #22 .

Change-Id: I0040b75671317ce355a32c4528159367835cb513