Introduce "Lock Strategy" option
With this commit, we allow using an in-application lock in ghostferry,
instead of using the source DB as lock. The lock is required to avoid
race conditions between the data iteration/copy and the binlog writer.

The default behavior is preserved; a new option LockStrategy allows
moving the lock from the source DB into ghostferry, or disabling the
lock altogether.

This fixes #24

Change-Id: I805b9490ab27e2a7674475d56ead68dd9b76c0e8