Fix handling of GRANT with resource-options
The SQL parser we use fails at parsing some GRANT statements:

https://github.com/pingcap/parser/issues/857

We don't need to support them, so this commit hacks around the problem
by detecting the error and ignoring it. This is ugly and must be
fixed... once we have additional cycles.

On the plus side: we add a few trivial (and desperately needed)
unit-tests for the query-analyzer.

Change-Id: I6b658720c030df6013f5093378e1c92a2837b07a