Better handling of corrupted resume state in DB
Improve error reporting when parsing the binlog writer resume state
from the target DB, and handle the fact when the state contains an
unknown table.
While the latter should never happen, we should still be more
resilient to unexpected resume state (e.g., if the table whitelist
changes across runs).

Increase unit-test coverage to a less-ridiculous level on the
state-in-DB parsing.

Change-Id: I55411289ac7e596a94bd716a151c17b5882b305a