Skip CREATE PROCEDURE/FUNCTION (#38)
Better (but still not good enough) handling of CREATE PROCEDURE and
CREATE FUNCTION statements. The SQL parser we use does not understand
these statements.

Starting with this commit, we log a warning but continue replication.
The statements are not core to replicating a DB content - although it
may not make the target an acceptable migration target after takeover!
Marking support out-of-scope / known-limitation for now.