Report Binlog Writer State (#41)
To improve monitoring of the ghostferry state, this commit introduces
two new metrics on the binlog writer:
- the "last state": if we are currently receiving events from the
  binlog server, are processing/applying them, or something else
- the last state-change timestamp: when the above state was last set.
  This allows detecting lockups (e.g., if we are stuck receiving updates
  from the binlog server).

The above metrics are integated in the status portal. Additionally we
introduce a new "health" API endpoint in the portal, which exports the
status as JSON.

In addition to the JSON export, we allow specifying a maximum age of the
last-change timestamp, returning an HTTP-500 if the age is exceeded.
This is convenient to use the health API as Kubernetes lifeness probe.