Skip DROP PROCEDURE/FUNCTION (#38)
Analogous to earlier commit for #38 but for DROP

Better (but still not good enough) handling of DROP PROCEDURE and
DROP FUNCTION statements. The SQL parser we use does not understand
these statements.

Starting with this commit, we log a warning but continue replication.
The statements are not core to replicating a DB content - although it
may not make the target an acceptable migration target after takeover!
Marking support out-of-scope / known-limitation for now.

Change-Id: I3812319d526cb4f910ffa95839e60a1e262f4054