Export TLSConfig to allow control of TLS to mysql
The source and destination db configs have a TLS option, but
it only supports setting a root ca and servername and not settings such as
InsecureSkipVerify. Internally the TLS config is used to construct a
standard tls.config struct which is passed to both the mysql and binlogstreamer libraries.
This change exposes that tls.config struct allowing it to be decoded from the config directly
For example
Source:
  TLS:
    TLSConfig:
      InsecureSkipVerify: true